First working version

+++ src/com/dollarsbank/model/Account.java
@@ -6,7 +6,7 @@ import java.util.List;
 public class Account {
 	
 	private int id;
-	private long balance;
+	private double balance;
 	private List<Transaction> transactions;
 	private String userId;
 	
@@ -15,30 +15,30 @@ public class Account {
 		this.transactions = new ArrayList<Transaction>();
 	}
 	
-	public Account(int id, long balance, String userId) {
+	public Account(int id, double balance, String userId) {
 		this.id = id;
 		this.balance = balance;
 		this.transactions = new ArrayList<Transaction>();
 		this.userId = userId;
 	}
 	
-	Account(long initBalance, String userId) {
+	Account(double initBalance, String userId) {
 		this.transactions = new ArrayList<Transaction>();
 		this.balance = initBalance;
 		this.transactions = new ArrayList<Transaction>();
 		this.userId = userId;
 	}
 	
-	public long getBalance() {
+	public double getBalance() {
 		return this.balance;
 	}
 	
-	public long withdraw(long amount) {
+	public double withdraw(double amount) {
 		this.balance -= amount;
 		return amount;
 	}
 	
-	public void deposit(long amount) {
+	public void deposit(double amount) {
 		this.balance += amount;
 	}
